bundle each folders test into 1 file
